todo에 접근할 때 jwt 검증

+++ server/routers/todo.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const { getConnection } = require("../models/connector");
+const jwt = require("jsonwebtoken");
 
 router.get("/", async (req, res) => {
   const [results] = await getConnection().execute(`SELECT * FROM todo`);
@@ -9,11 +10,18 @@ router.get("/", async (req, res) => {
 
 router.post("/", async (req, res) => {
   const data = req.body;
-  await getConnection().execute(
-    `INSERT INTO todo (todo, completed) VALUES (?,?)`,
-    [data.todo, data.completed]
-  );
-  return res.json("success");
+  const token = req.headers.authorization;
+  try {
+    const tokenResult = jwt.verify(token, "secret");
+    await getConnection().execute(
+      `INSERT INTO todo (todo, completed, authorId) VALUES (?,?,?)`,
+      [data.todo, data.completed, data.authorId]
+    );
+    return res.json("success");
+  } catch (error) {
+    console.error(error);
+    return res.status(403).json("invalid token");
+  }
 });
 
 router.put("/:id", async (req, res) => {
@@ -28,8 +36,27 @@ router.put("/:id", async (req, res) => {
 
 router.delete("/:id", async (req, res) => {
   const id = req.params.id;
-  await getConnection().execute(`DELETE FROM todo WHERE id =?`, [id]);
-  return res.json("success");
+  const token = req.headers.authorization;
+
+  try {
+    const tokenResult = jwt.verify(token, "secret");
+    const currentId = tokenResult.id;
+
+    const [results] = await getConnection().execute(
+      `SELECT * FROM todo WHERE id =?`,
+      [id]
+    );
+
+    if (results[0].authorId !== currentId) {
+      return res.status(401).json("no auth");
+    }
+
+    await getConnection().execute(`DELETE FROM todo WHERE id =?`, [id]);
+    return res.json("success");
+  } catch (error) {
+    console.error(error);
+    return res.status(403).json("invalid token");
+  }
 });
 
 module.exports = router;
